Fix: useSearchParams build prerendering error
