--- src/app/quiz/layout.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-type Props = {
-    children: React.ReactNode;
-};
-
-const QuizLayout = ({ children }: Props) => <div>{children}</div>;
-
-export default QuizLayout;
