+++ src/components/screens/quiz/QuizSwiper.tsx
@@ -1,3 +1,4 @@
+'use client';
 import React, { FC, useEffect, useRef, useState } from 'react';
 import { getRandomQuizList } from '@src/lib/quiz/getRandomQuizList';
 import { QuizListService } from '@src/service/QuizListService';
